first page + fixes

+++ src/main/java/ru/luckyboy/aggregator/repository/NewsItemRepository.java
@@ -3,7 +3,7 @@ package ru.luckyboy.aggregator.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.luckyboy.aggregator.domain.NewsItem;
 
-public interface NewsItemRepository extends JpaRepository<NewsItem, Integer> {
+public interface NewsItemRepository extends JpaRepository<NewsItem, Long> {
 
     NewsItem findByLink(String link);
 }
