+++ src/main/java/ru/luckyboy/aggregator/repository/ParseRuleRepository.java
@@ -3,5 +3,5 @@ package ru.luckyboy.aggregator.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.luckyboy.aggregator.domain.ParseRule;
 
-public interface ParseRuleRepository extends JpaRepository<ParseRule, Integer> {
+public interface ParseRuleRepository extends JpaRepository<ParseRule, Long> {
 }
